+++ .eslintrc.js
@@ -1,0 +1,41 @@
+module.exports = {
+  plugins: ['import', 'unused-imports', 'react'],
+  rules: {
+    'react/jsx-uses-react': 'error',
+    'react/jsx-uses-vars': 'error',
+    'unused-imports/no-unused-imports': 'error',
+  },
+  parserOptions: {
+    sourceType: "module",
+    ecmaVersion: 11,
+    ecmaFeatures: {
+      jsx: true,
+    }
+  },
+  overrides: [
+    {
+      files: ['frontend/src/**/*.{js,jsx}'],
+      rules: {
+        'sort/imports': 0,
+        'import/order': [
+          'error',
+          {
+            groups: [
+              'builtin',
+              'external',
+              'parent',
+              'sibling',
+              'index',
+              'object',
+              'type',
+            ],
+            alphabetize: {
+              order: 'asc',
+            },
+            'newlines-between': 'always',
+          },
+        ],
+      },
+    },
+  ],
+};