+++ airone/urls.py
@@ -1,12 +1,12 @@
 from django.conf import settings
 from django.conf.urls import url, include
-from django.contrib import admin
 from django.views.generic import RedirectView
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 from django.contrib.auth import views as auth_views
 from . import views
 
 from api_v1.urls import urlpatterns as api_v1_urlpatterns
+from airone.auth import view as auth_view
 
 urlpatterns = [
     url('^$', views.index, name='index'),
@@ -16,17 +16,17 @@ urlpatterns = [
     url(r'^group/', include(('group.urls', 'group'))),
     url(r'^entity/', include(('entity.urls', 'entity'))),
     url(r'^dashboard/', include(('dashboard.urls', 'dashboard'))),
+    url(r'^new-ui/', include(('dashboard.urls_for_new_ui', 'dashboard_for_new_ui'))),
     url(r'^entry/', include(('entry.urls', 'entry'))),
-    url(r'^admin/', admin.site.urls),
-    url(r'^api-auth/', include(('rest_framework.urls', 'rest_framework'))),
     url(r'^api/v1/', include(api_v1_urlpatterns)),
     url(r'^job/', include(('job.urls', 'job'))),
-    url(r'^auth/', include(('django.contrib.auth.urls', 'auth'))),
-    url(r'^webhook/', include(('webhook.urls', 'entity'))),
-    url(r'^login/', auth_views.LoginView.as_view(), name='login'),
-    url(r'^logout/', auth_views.LogoutView.as_view(), name='logout'),
+    url(r'^auth/login/', auth_views.LoginView.as_view(
+        redirect_authenticated_user=True,
+    ), name='login'),
+    url(r'^auth/logout/', auth_view.logout, name='logout'),
     url(r'^oauth/', include('social_django.urls', namespace='social')),
     url(r'^accounts/profile/', views.index, name='index'),
+    url(r'^webhook/', include(('webhook.urls', 'webhook'))),
 ]
 
 for extension in settings.AIRONE['EXTENSIONS']: