+++ frontend/src/components/group/GroupList.test.js
@@ -1,0 +1,12 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import { render } from "@testing-library/react";
+import React from "react";
+
+import { GroupList } from "./GroupList";
+
+test("should render a component with essential props", function () {
+  expect(() => render(<GroupList groups={[]} />)).not.toThrow();
+});