+++ user/urls.py
@@ -1,11 +1,11 @@
-from django.conf.urls import url
+from django.conf.urls import url, include
 
-from airone.lib.log import Logger
 from group import views as group_views
 from user import views
 
 urlpatterns = [
     url(r'^$', views.index, name='index'),
+    url(r'^api/v2/', include(('user.api_v2.urls', 'user.api_v2'))),
     url(r'^edit/(\d+)$', views.edit, name='edit'),
     url(r'^do_edit/(\d+)$', views.do_edit, name='do_edit'),
     url(r'^edit_passwd/(\d+)$', views.edit_passwd, name='edit_passwd'),
@@ -14,6 +14,7 @@ urlpatterns = [
     url(r'^create$', views.create, name='create'),
     url(r'^do_create$', views.do_create, name='do_create'),
     url(r'^do_delete/(\d+)$', views.do_delete, name='do_delete'),
+    url(r'^change_ldap_auth$', views.change_ldap_auth, name='change_ldap_auth'),
     url(r'^export/$', group_views.export, name='export'),
     url(r'^password_reset/$', views.PasswordReset.as_view(), name='password_reset'),
     url(r'^password_reset/done/$', views.PasswordResetDone.as_view(), name='password_reset_done'),
@@ -21,10 +22,3 @@ urlpatterns = [
         views.PasswordResetConfirm.as_view(), name='password_reset_confirm'),
     url(r'^reset/done/$', views.PasswordResetComplete.as_view(), name='password_reset_complete'),
 ]
-
-try:
-    from custom_view.user.urls import override_urlpatterns
-
-    urlpatterns = override_urlpatterns(urlpatterns)
-except ImportError:
-    Logger.info("There is no URL dispatcher of custom-view")